requirements.txt
# Copyright (c) 2017-2021 VMware, Inc. All Rights Reserved.
# SPDX-License-Identifier: BSD-2-Clause
#
# Please only add direct dependencies here, i.e., do not update with the
# output of `pip freeze`.  When updating dependency versions, having the
# transitive dependencies listed make it more difficult to figure out
# what should be updated.

PyYAML>=5.4
docker~=5.0
dockerfile-parse~=1.2
requests~=2.26
stevedore>=3.4
pbr>=5.6
debian-inspector>=30.0
regex>=2021.9
GitPython~=3.1
prettytable~=2.2
packageurl-python>=0.9.4
